Update exrtma3d_clean.ksh
Added additional conditional blocks that check for the existence of COM and DATA directories before removing individual files coinciding with the retention period.

+++ scripts/exrtma3d_clean.ksh
@@ -1,19 +1,5 @@
 #!/bin/ksh
 
-#
-#-----------------------------------------------------------------------
-# Source the variable definitions file.
-#-----------------------------------------------------------------------
-#
-#
-#-----------------------------------------------------------------------
-# Save current shell options (in a global array).  Then set new options
-# for this script/function.
-#-----------------------------------------------------------------------
-#
-#
-#-----------------------------------------------------------------------
-# set up currentime from CDATE 
 #-----------------------------------------------------------------------
 #
 currentime=$(echo "${PDYHH}" | sed 's/\([[:digit:]]\{2\}\)$/ \1/')
@@ -28,9 +14,16 @@ set -A XX $(ls -d ${NET}.20* | sort -r)
 for dir in ${XX[*]};do
   onetime=$(echo $dir | cut -d'.' -f2)
   if [[ ${onetime} =~ ^[0-9]+$ ]] && [[ ${onetime} -le ${deletetime} ]]; then
-    rm -rf ${COMROOT}/${NET}/pbspro/${NET}.${onetime}
-    echo "Deleted ${COMROOT}/${NET}/pbspro/${NET}.${onetime}"
-  fi
+    if [ -d ${COMROOT}/${NET}/pbspro/${NET}.${onetime} ]; then
+      rm -rf ${COMROOT}/${NET}/pbspro/${NET}.${onetime}/*t${cyc}*
+      echo "Deleted t${cyc}z files in ${COMROOT}/${NET}/pbspro/${NET}.${onetime}"
+      if [ ${cyc} == 23 ]; then
+            rm -rf ${COMROOT}/${NET}/pbspro/${NET}.${onetime}
+            echo "Deleted ${COMROOT}/${NET}/pbspro/${NET}.${onetime}"
+      fi
+    else
+      echo "Nothing to delete."
+    fi
 done
 
 #-----------------------------------------------------------------------
@@ -43,8 +36,12 @@ set -A XX $(ls -d ${NET}.20* | sort -r)
 for dir in ${XX[*]};do
   onetime=$(echo $dir | cut -d'.' -f2)
   if [[ ${onetime} =~ ^[0-9]+$ ]] && [[ ${onetime} -le ${deletetime} ]]; then
-    rm -rf ${DATAROOT}/pbspro/${NET}.${onetime}
-    echo "Deleted ${DATAROOT}/pbspro/${NET}.${onetime}"
+    if [ -d ${DATAROOT}/pbspro/${NET}.${onetime} ]; then
+      rm -rf ${DATAROOT}/pbspro/${NET}.${onetime}
+      echo "Deleted ${DATAROOT}/pbspro/${NET}.${onetime}"
+    else
+      echo "Nothing to delete."
+    fi
   fi
 done
